pyproject.toml
[project]
name = "personal-intern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.39.0",
    "httpx==0.27.2",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "scrapybara>=0.2.2",
    "telethon>=1.38.1",
    "typer>=0.15.1",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
